moved java files into device and display packages
