+++ pom.xml
@@ -44,8 +44,12 @@
       <artifactId>angularjs</artifactId>
       <version>1.2.26</version>
     </dependency>
-
-
+    <dependency>
+      <groupId>org.glassfish.jersey.containers</groupId>
+      <artifactId>jersey-container-jetty-servlet</artifactId>
+      <version>2.13</version>
+    </dependency>
+    
     <!-- Compile/runtime dependencies -->
     <dependency>
       <groupId>com.google.appengine</groupId>
