+++ src/main/java/com/openhouseautomation/devices/LocationServlet.java
@@ -1,4 +1,4 @@
-package com.openhouseautomation;
+package com.openhouseautomation.devices;
 
 import java.io.IOException;
 import java.io.PrintWriter;
